# perf(drm): group dracut_instmods calls